+++ AA/AA-CHMU-SMS-ORP.js
@@ -1,5 +1,5 @@
 //!JS
-// Verze 21
+// Verze 22
 
 // zde např. Mělník. Číselník ORP viz samostatný soubor
 var omezitNaOrp = 141; 
@@ -10,9 +10,12 @@ var oddelovac = '\n';
 // zde vytvoříme tělo SMS dle obsahu CAP pomocí skriptu z knihovny
 #import "CHMU-SMS-ORP"; 
 
-// Upozorňuji, že tělo SMS zprávy nekončí zalomením řádku. Já si proto odřádkuji. Ale vy nemusíte, pokud nechcete.
-vystupText += "\n";
+if (vystupText != '') {
+    // Upozorňuji, že tělo SMS zprávy nekončí zalomením řádku. Já si proto odřádkuji. Ale vy nemusíte, pokud nechcete.
+    vystupText += "\n";
 
-// Podpis na konci každé SMS.
-vystupText += "OPIS GŘ HZS ČR";
+    // Podpis na konci každé SMS.
+    vystupText += "OPIS GŘ HZS ČR";
+}
+
 print(vystupText);